voice_assistant_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.26.2
scipy==1.11.4
soundfile==0.12.1
datasets==2.15.0
sentencepiece==0.1.99
arxiv==1.4.8
sympy==1.12
